Rename sound.rs to audio.rs and split audio generator and audio processor to module
